Created daily email report
