+++ .gitignore
@@ -2,3 +2,4 @@
 .webassets-cache
 .git
 .idea/
+*_test.py
