+++ app.py
@@ -1,8 +1,9 @@
 from flask import Flask, request, redirect, render_template
-from TextChella import is_user, subscribe, delete_user
+from text_chella import is_user, subscribe, delete_user
 from twilio.twiml.messaging_response import MessagingResponse
 
 app = Flask(__name__)
+
 
 @app.route("/textChella", methods=['POST', 'GET'])
 def response():
@@ -24,8 +25,6 @@ def response():
         else:
             return ''
 
-
-    #test
 
 if __name__ == "__main__":
     app.run(debug=True)
