+++ clock.py
@@ -1,10 +1,12 @@
-from TextChella import text_chella
+from text_chella import text_chella
 from apscheduler.schedulers.blocking import BlockingScheduler
 
 sched = BlockingScheduler()
 
+
 @sched.scheduled_job('cron', hour=10, timezone='US/Pacific')
 def send_text():
     text_chella()
-    
+
+
 sched.start()
